re_dataframe_ui::display_record_batch
Function quick_partial_hash

fn quick_partial_hash(data: &[u8], section_length: usize) -> Hash64

Compute a quick partial hash of an image data buffer, capping the max amount of hashed data to 3*section_length.

If the buffer is smaller or equal to than 3*section_length, the entire buffer is hashed. If the buffer is larger, the first, middle, and last sections, each of size section_length, are hashed.