Crate re_dataframe_ui

Rich table widget over datafusion.

Modules

datafusion_adapter 🔒

datafusion_table_widget 🔒

display_record_batch 🔒

Intermediate data structures to make re_datastore’s row data more amenable to displaying in a table.

requested_object 🔒

table_blueprint 🔒

table_utils

Structs

DataFusionTableWidget

DisplayRecordBatch

Enums

DisplayRecordBatchError

RequestedObject

A handle to an object that is requested asynchronously.